requirements.txt
django>=4.2.19
google-generativeai>=0.3.2
python-dotenv>=1.0.0
Pillow>=10.0.0 
